requirements.txt
matplotlib>=3.9
pandas>=2.2
openpyxl>=3.1
transformers>=4.44
python-dotenv>=1.0
